+++ src/connection.js
@@ -1,5 +1,6 @@
 const Environment = require('./env');
 const Utils = require('@openeo/js-commons/src/utils');
+const ProcessRegistry = require('@openeo/js-commons/src/processRegistry');
 const axios = require('axios').default;
 const StacMigrate = require('@radiantearth/stac-migrate');
 
@@ -25,27 +26,60 @@ class Connection {
 	/**
 	 * Creates a new Connection.
 	 * 
-	 * @param {string} baseUrl - URL to the back-end
+	 * @param {string} baseUrl - URL to the back-end.
+	 * @param {Options} [options={}] - Additional options for the connection.
 	 */
-	constructor(baseUrl) {
+	constructor(baseUrl, options = {}) {
 		/**
+		 * URL of the backend connected to.
+		 * 
+		 * @protected
 		 * @type {string}
 		 */
 		this.baseUrl = Utils.normalizeUrl(baseUrl);
 		/**
+		 * Auth Provider cache
+		 * 
+		 * @protected
 		 * @type {?Array.<AuthProvider>}
 		 */
 		this.authProviderList = null;
 		/**
+		 * Current auth provider
+		 * 
+		 * @protected
 		 * @type {?AuthProvider}
 		 */
 		this.authProvider = null;
 		/**
+		 * Capability cache
+		 * 
+		 * @protected
 		 * @type {?Capabilities}
 		 */
 		this.capabilitiesObject = null;
-		this.processes = null;
+		/**
+		 * Process cache
+		 * 
+		 * @protected
+		 * @type {ProcessRegistry}
+		 */
+		this.processes = new ProcessRegistry([], Boolean(options.addNamespaceToProcess));
+		this.processes.listeners.push((...args) => this.emit('processesChanged', ...args));
+		/**
+		 * Listeners for events.
+		 * 
+		 * @protected
+		 * @type {object.<string|Function>}
+		 */
 		this.listeners = {};
+		/**
+		 * Additional options for the connection.
+		 * 
+		 * @protected
+		 * @type {Options}
+		 */
+		this.options = options;
 	}
 
 	/**
@@ -156,16 +190,16 @@ class Connection {
 	 * 
 	 * @async
 	 * @param {string} collectionId - Collection ID to request items for.
-	 * @param {?Array.<number>} spatialExtent - Limits the items to the given bounding box in WGS84:
+	 * @param {?Array.<number>} [spatialExtent=null] - Limits the items to the given bounding box in WGS84:
 	 * 1. Lower left corner, coordinate axis 1
 	 * 2. Lower left corner, coordinate axis 2
 	 * 3. Upper right corner, coordinate axis 1
 	 * 4. Upper right corner, coordinate axis 2
-	 * @param {?Array.<*>} temporalExtent - Limits the items to the specified temporal interval.
+	 * @param {?Array.<*>} [temporalExtent=null] - Limits the items to the specified temporal interval.
 	 * The interval has to be specified as an array with exactly two elements (start, end) and
 	 * each must be either an RFC 3339 compatible string or a Date object.
 	 * Also supports open intervals by setting one of the boundaries to `null`, but never both.
-	 * @param {?number} limit - The amount of items per request/page as integer. If `null` (default), the back-end decides.
+	 * @param {?number} [limit=null] - The amount of items per request/page as integer. If `null` (default), the back-end decides.
 	 * @yields {Promise<ItemCollection>} A response compatible to the API specification.
 	 * @throws {Error}
 	 */
@@ -208,20 +242,35 @@ class Connection {
 	}
 
 	/**
-	 * List all processes available on the back-end.
+	 * List processes available on the back-end.
 	 * 
-	 * Data is cached in memory.
+	 * Requests pre-defined processes by default.
+	 * Set the namespace parameter to request processes from a specific namespace.
+	 * 
+     * Note: The list of namespaces can be retrieved by calling `listProcesses` without a namespace given.
+	 * The namespaces are then listed in the property `namespaces`.
 	 * 
 	 * @async
+	 * @param {?string} [namespace=null] - Namespace of the processes (default to `null`, i.e. pre-defined processes). EXPERIMENTAL!
 	 * @returns {Promise<Processes>} - A response compatible to the API specification.
 	 * @throws {Error}
 	 */
-	async listProcesses() {
-		if (this.processes === null) {
-			let response = await this._get('/processes');
-			this.processes = response.data;
+	async listProcesses(namespace = null) {
+		if (!namespace) {
+			namespace = 'backend';
 		}
-		return this.processes;
+		let path = (namespace === 'backend') ? '/processes' : `/processes/${namespace}`;
+		let response = await this._get(path);
+
+		if (!Utils.isObject(response.data) || !Array.isArray(response.data.processes)) {
+			throw new Error('Invalid response received for processes');
+		}
+
+		// Store processes in cache
+		this.processes.remove(null, namespace);
+		this.processes.addAll(response.data.processes, namespace);
+		
+		return Object.assign(response.data, {processes: this.processes.namespace(namespace)});
 	}
 
 	/**
@@ -229,20 +278,30 @@ class Connection {
 	 * 
 	 * @async
 	 * @param {string} processId - Collection ID to request further metadata for.
+	 * @param {?string} [namespace=null] - Namespace of the process (default to `null`, i.e. pre-defined processes). EXPERIMENTAL!
 	 * @returns {Promise<?Process>} - A single process as object, or `null` if none is found.
 	 * @throws {Error}
 	 * @see Connection#listProcesses
 	 */
-	async describeProcess(processId) {
-		let response = await this.listProcesses();
-		if (Array.isArray(response.processes)) {
-			return response.processes.find(process => Utils.isObject(process) && process.id === processId) || null;
+	async describeProcess(processId, namespace = null) {
+		if (!namespace) {
+			namespace = 'backend';
 		}
-		return null;
+		if (namespace === 'backend') {
+			await this.listProcesses();
+		}
+		else {
+			let response = await this._get(`/processes/${namespace}/${processId}`);
+			if (!Utils.isObject(response.data) || typeof response.data.id !== 'string') {
+				throw new Error('Invalid response received for process');
+			}
+			this.processes.add(response.data, namespace);
+		}
+		return this.processes.get(processId, namespace);
 	}
 
 	/**
-	 * Returns an object to simply build user-defined processes.
+	 * Returns an object to simply build user-defined processes based upon pre-defined processes.
 	 * 
 	 * @async
 	 * @param {string} id - A name for the process.
@@ -251,8 +310,8 @@ class Connection {
 	 * @see Connection#listProcesses
 	 */
 	async buildProcess(id) {
-		let response = await this.listProcesses();
-		return new Builder(response.processes, null, id);
+		await this.listProcesses();
+		return new Builder(this.processes, null, id);
 	}
 
 	/**
@@ -305,7 +364,7 @@ class Connection {
 	 *
 	 * @callback oidcProviderFactoryFunction
 	 * @param {object.<string, *>} providerInfo - The provider information as provided by the API, having the properties `id`, `issuer`, `title` etc.
-	 * @returns {?AuthProvider}
+	 * @returns {AuthProvider | null}
 	 */
 
 	/**
@@ -315,7 +374,7 @@ class Connection {
 	 * on the AuthProvider interface (or OIDCProvider class), e.g. to use a
 	 * OIDC library other than oidc-client-js.
 	 * 
-	 * @param {?oidcProviderFactoryFunction} providerFactoryFunc
+	 * @param {?oidcProviderFactoryFunction} [providerFactoryFunc=null]
 	 * @see AuthProvider
 	 */
 	setOidcProviderFactory(providerFactoryFunc) {
@@ -328,7 +387,7 @@ class Connection {
 	 * Returns `null` if OIDC is not supported by the client or an instance
 	 * can't be created for whatever reason.
 	 * 
-	 * @returns {?oidcProviderFactoryFunction}
+	 * @returns {oidcProviderFactoryFunction | null}
 	 * @see AuthProvider
 	 */
 	getOidcProviderFactory() {
@@ -390,6 +449,7 @@ class Connection {
 	 * Currently supported:
 	 * - authProviderChanged(provider): Raised when the auth provider has changed.
 	 * - tokenChanged(token): Raised when the access token has changed.
+	 * - processesChanged(type, data, namespace): Raised when the process registry has changed (i.e. a process was added, updated or deleted)
 	 * 
 	 * @param {string} event 
 	 * @param {Function} callback 
@@ -410,7 +470,7 @@ class Connection {
 	/**
 	 * Returns the AuthProvider.
 	 * 
-	 * @returns {?AuthProvider} 
+	 * @returns {AuthProvider | null} 
 	 */
 	getAuthProvider() {
 		return this.authProvider;
@@ -582,6 +642,15 @@ class Connection {
 	 */
 	async listUserProcesses() {
 		let response = await this._get('/process_graphs');
+
+		if (!Utils.isObject(response.data) || !Array.isArray(response.data.processes)) {
+			throw new Error('Invalid response received for processes');
+		}
+
+		// Store processes in cache
+		this.processes.remove(null, 'user');
+		this.processes.addAll(response.data.processes, 'user');
+
 		return response.data.processes.map(
 			pg => new UserProcess(this, pg.id).setAll(pg)
 		);
@@ -824,7 +893,7 @@ class Connection {
 	 * 
 	 * @param {Array.<Link>} links - An array of links.
 	 * @param {string} rel - Relation type to find, defaults to `next`.
-	 * @returns {?string}
+	 * @returns {string | null}
 	 * @throws {Error}
 	 */
 	_getLinkHref(links, rel = 'next') {
@@ -855,7 +924,7 @@ class Connection {
 			url: path,
 			// Timeout for capabilities requests as they are used for a quick first discovery to check whether the server is a openEO back-end.
 			// Without timeout connecting with a wrong server url may take forever.
-			timeout: path === '/' ? 3000 : 0,
+			timeout: path === '/' ? 5000 : 0,
 			params: query
 		});
 	}