+++ tests/basic.test.js
@@ -1,9 +1,15 @@
 // @ts-nocheck
-const { OpenEO } = require('../src/openeo');
+const { OpenEO, OidcProvider } = require('../src/openeo');
 const packageInfo = require('../package.json');
 
 describe('Client Basics', () => {
 	test('Check version number', () => {
 		expect(OpenEO.clientVersion()).toBe(packageInfo.version);
+	});
+
+	describe('OIDC', () => {
+		test('isSupported', async () => {
+			expect(OidcProvider.isSupported()).toBeTruthy();
+		});
 	});
 });